Add property declarations to PhabObject
This lets us add consistent behavior among initialize, create, update, and
property accessor methods.

As an initial set of property options, this PR adds the ability to
automatically translate names to raw values (usually PHIDs, but can
also be things like priority numbers).

As a final bonus, this includes the first ever tests for this repo. :)